Detects when it is not pegin by unsupported sender address and does not use pegin version 1.
If a transaction to the powpeg address is sent by an unsupported address, and pegin v1 (OP_RETURN) is not used then it considers that it is not a pegin.

In addition, it is resolved as the RSK address is shown when it is obtained from OP_RETURN, adding 0x to it.