Resolve bug when the tx is from Native Segwit or Segwit Compatibility address.
When the address contains witness, the hash obtained from the RSK nodes includes witness, but the has used an entry parameter doesn't contains it. This fix verify the tx is from this kind of addresses, obtains the hash of a hex parametrer of the transaction and ask to the database using this hash.